# spec/api_contract.yaml
openapi: 3.0.3
info:
  title: ChronicleCore Local API
  description: |
    Local-only REST API for ChronicleCore time tracking application.
    All endpoints bind to 127.0.0.1 only. No remote access permitted.
  version: 1.0.0
  contact:
    name: ChronicleCore
servers:
  - url: http://127.0.0.1:{port}
    description: Local development server
    variables:
      port:
        default: '8080'
        description: Configurable port (default 8080)
...
paths:
...
  /api/v1/profiles:
    get:
      summary: List profiles
      tags: [Profiles]
      parameters:
        - name: active_only
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Profile'

    post:
      summary: Create profile
      tags: [Profiles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileCreate'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/profiles/{profile_id}:
    get:
      summary: Get profile
      tags: [Profiles]
      parameters:
        - name: profile_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Profile details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update profile
      tags: [Profiles]
      parameters:
        - name: profile_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileCreate'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete profile
      tags: [Profiles]
      parameters:
        - name: profile_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Profile deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  schemas:
...
    Profile:
      type: object
      properties:
        profile_id:
          type: integer
        client_name:
          type: string
        project_name:
          type: string
          nullable: true
        service_name:
          type: string
        rate_name:
          type: string
        rate_amount:
          type: number
          description: Hourly rate in major units
        currency:
          type: string
        is_active:
          type: boolean
    ProfileCreate:
      type: object
      required: [client_id, service_id, rate_id]
      properties:
        client_id:
          type: integer
        project_id:
          type: integer
          nullable: true
        service_id:
          type: integer
        rate_id:
          type: integer
        name:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - NOT_FOUND
                - INTERNAL_ERROR
                - INVALID_STATE
            message:
              type: string
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INVALID_REQUEST
              message: Invalid parameter value
              details: {}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
              details: {}
security: []